Add Generate PDF Report functionality
This code adds a button to the viewer window titled "Generate PDF". Pressing the button will generate a PDF report of all areas of interest identified in images.
