+++ app/core/controllers/Viewer.py
@@ -14,18 +14,7 @@ from core.services.XmlService import XmlService
 from helpers.LocationInfo import LocationInfo
 from core.services.LoggerService import LoggerService
 from helpers.MetaDataHelper import MetaDataHelper
-from reportlab.lib import colors
-from reportlab.lib.pagesizes import letter
-from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
-from reportlab.lib.units import inch
-from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Image, Table, TableStyle
-from reportlab.platypus import PageBreak
-from reportlab.platypus.tableofcontents import TableOfContents
-from reportlab.platypus.doctemplate import PageTemplate, BaseDocTemplate
-from reportlab.platypus.frames import Frame
-from reportlab.lib.units import cm
-import tempfile
-import os
+from core.services.PdfGeneratorService import PdfGeneratorService
 
 
 class Viewer(QMainWindow, Ui_Viewer):
@@ -583,195 +572,12 @@ class Viewer(QMainWindow, Ui_Viewer):
         """Handles clicks on the Generate PDF button."""
         fileName, _ = QFileDialog.getSaveFileName(self, "Save PDF File", "", "PDF files (*.pdf)")
         if fileName:
-            self.generatePdf(fileName)
-
-    def generatePdf(self, output_path):
-        """Generates a PDF report of the analysis results.
-
-        Args:
-            output_path (str): The file path to save the PDF file.
-        """
-        try:
-            # Create document template
-            doc = PDFDocTemplate(output_path, pagesize=letter)
-            styles = getSampleStyleSheet()
-            story = []
-            temp_files = []  # Keep track of temp files
-
-            # Define heading styles
-            h1 = ParagraphStyle(
-                name='Heading1',
-                parent=styles['Heading1'],
-                fontSize=24,
-                spaceAfter=30
-            )
-            
-            h2 = ParagraphStyle(
-                name='Heading2',
-                parent=styles['Heading2'],
-                fontSize=18,
-                spaceAfter=20
-            )
-
-            # Create Table of Contents
-            toc = TableOfContents()
-            toc.levelStyles = [
-                ParagraphStyle(
-                    fontSize=14, 
-                    name='TOCHeading1',
-                    leftIndent=20, 
-                    firstLineIndent=-20, 
-                    spaceBefore=5, 
-                    leading=16
-                ),
-                ParagraphStyle(
-                    fontSize=12, 
-                    name='TOCHeading2',
-                    leftIndent=40, 
-                    firstLineIndent=-20, 
-                    spaceBefore=0, 
-                    leading=12
-                )
-            ]
-            
-            # Add title and TOC
-            story.append(Paragraph("Analysis Report", h1))
-            story.append(toc)
-            story.append(PageBreak())
-
-            # Algorithm Information
-            story.append(Paragraph("Algorithm Settings", h1))
-            settings, algorithm = self.xmlService.getSettings()
-            story.append(Paragraph(f"Algorithm: {algorithm}", styles['Normal']))
-            
-            # Process settings with color squares
-            for key, value in settings.items():
-                rgb_value = value if isinstance(value, tuple) else parse_rgb_string(str(value))
-                
-                if rgb_value and len(rgb_value) == 3:
-                    r, g, b = rgb_value
-                    color_hex = f"#{r:02x}{g:02x}{b:02x}"
-                    story.append(Paragraph(
-                        f"{key}: {value} <font color='{color_hex}'>■</font>",
-                        styles['Normal']
-                    ))
-                elif isinstance(value, dict):
-                    nested_values = []
-                    for k, v in value.items():
-                        rgb_value = parse_rgb_string(str(v))
-                        if rgb_value and len(rgb_value) == 3:
-                            r, g, b = rgb_value
-                            color_hex = f"#{r:02x}{g:02x}{b:02x}"
-                            nested_values.append(f"{k}: {v} <font color='{color_hex}'>■</font>")
-                        else:
-                            nested_values.append(f"{k}: {v}")
-                    story.append(Paragraph(f"{key}: {{{', '.join(nested_values)}}}", styles['Normal']))
-                else:
-                    story.append(Paragraph(f"{key}: {value}", styles['Normal']))
-
-            story.append(Spacer(1, 20))
-
-            # Total Areas of Interest
-            total_aoi = sum(len(img['areas_of_interest']) for img in self.images)
-            story.append(Paragraph(f"Total Areas of Interest: {total_aoi}", h1))
-            story.append(PageBreak())
-
-            # Image Details
-            for img in self.images:
-                if len(img['areas_of_interest']) > 0:
-                    # Image Header (will be added to TOC)
-                    story.append(Paragraph(f"Image: {img['name']}", h1))
-                    
-                    # GPS Coordinates
-                    gps_coords = LocationInfo.getGPS(img['path'])
-                    if gps_coords:
-                        position = self.getPosition(gps_coords['latitude'], gps_coords['longitude'])
-                        story.append(Paragraph(f"GPS Coordinates: {position}", styles['Normal']))
-                    story.append(Spacer(1, 20))
-
-                    # Add full image
-                    temp_file = tempfile.NamedTemporaryFile(suffix='.png', delete=False)
-                    temp_files.append(temp_file.name)
-                    temp_file.close()
-
-                    # Save full image
-                    img_arr = qimage2ndarray.imread(img['path'])
-                    qimg = qimage2ndarray.array2qimage(img_arr)
-                    scaled_img = qimg.scaled(1500, 1500, Qt.KeepAspectRatio, Qt.SmoothTransformation)
-                    scaled_img.save(temp_file.name)
-
-                    # Add to PDF with proper sizing
-                    img_obj = Image(temp_file.name)
-                    available_width = doc.width
-                    available_height = doc.height * 0.5
-                    img_width = img_obj.imageWidth
-                    img_height = img_obj.imageHeight
-                    
-                    scale = min(available_width / img_width, available_height / img_height)
-                    img_obj.drawWidth = img_width * scale
-                    img_obj.drawHeight = img_height * scale
-                    story.append(img_obj)
-                    story.append(Spacer(1, 20))
-
-                    # Areas of Interest
-                    for i, aoi in enumerate(img['areas_of_interest'], 1):
-                        story.append(Paragraph(f"Area of Interest #{i}", h2))
-                        story.append(Paragraph(
-                            f"Pixel Coordinates: ({aoi['center'][0]}, {aoi['center'][1]})", 
-                            styles['Normal']
-                        ))
-
-                        temp_file = tempfile.NamedTemporaryFile(suffix='.png', delete=False)
-                        temp_files.append(temp_file.name)
-                        temp_file.close()
-
-                        self.saveAoiThumbnail(img, aoi, temp_file.name)
-                        img_obj = Image(temp_file.name, width=3*inch, height=3*inch)
-                        story.append(img_obj)
-                        story.append(Spacer(1, 20))
-
-                    story.append(PageBreak())
-
-            # Build the PDF (using multiBuild for TOC)
-            doc.multiBuild(story)
-
-            # Clean up temporary files
-            for temp_file in temp_files:
-                try:
-                    os.unlink(temp_file)
-                except Exception as e:
-                    self.logger.error(f"Error deleting temporary file {temp_file}: {str(e)}")
-
-        except Exception as e:
-            self.logger.error(f"Error generating PDF file: {str(e)}")
-            self.showError("Failed to generate PDF file.")
-            # Clean up temporary files if there was an error
-            for temp_file in temp_files:
-                try:
-                    os.unlink(temp_file)
-                except:
-                    pass
-
-    def saveAoiThumbnail(self, image, area_of_interest, output_path):
-        """Saves an area of interest thumbnail to a file.
-
-        Args:
-            image (dict): The image data dictionary.
-            area_of_interest (dict): The area of interest data dictionary.
-            output_path (str): Path to save the thumbnail.
-        """
-        img_arr = qimage2ndarray.imread(image['path'])
-        center = area_of_interest['center']
-        radius = area_of_interest['radius'] + 10
-        crop_arr = self.crop_image(
-            img_arr, 
-            center[0] - radius, 
-            center[1] - radius, 
-            center[0] + radius, 
-            center[1] + radius
-        )
-        img = qimage2ndarray.array2qimage(crop_arr)
-        img.save(output_path)
+            try:
+                pdf_generator = PdfGeneratorService(self)
+                pdf_generator.generate_report(fileName)
+            except Exception as e:
+                self.logger.error(f"Error generating PDF file: {str(e)}")
+                self.showError(f"Failed to generate PDF file: {str(e)}")
 
 
 class ThumbnailLoader(QThread):
@@ -805,48 +611,3 @@ class ThumbnailLoader(QThread):
                 pixmap = QPixmap.fromImage(reader.read())
                 icon = QIcon(pixmap)
                 self.thumbnailLoaded.emit(index, icon)
-
-
-class PDFDocTemplate(BaseDocTemplate):
-    """Custom document template with TOC support."""
-    
-    def __init__(self, filename, **kwargs):
-        self.allowSplitting = 0
-        BaseDocTemplate.__init__(self, filename, **kwargs)
-        template = PageTemplate('normal', [Frame(2.5*cm, 2.5*cm, 15*cm, 25*cm, id='F1')])
-        self.addPageTemplates(template)
-
-    def afterFlowable(self, flowable):
-        """Registers TOC entries."""
-        if flowable.__class__.__name__ == 'Paragraph':
-            text = flowable.getPlainText()
-            style = flowable.style.name
-            if style == 'Heading1':
-                key = f'h1-{text}'
-                self.canv.bookmarkPage(key)
-                self.notify('TOCEntry', (0, text, self.page, key))
-            elif style == 'Heading2':
-                key = f'h2-{text}'
-                self.canv.bookmarkPage(key)
-                self.notify('TOCEntry', (1, text, self.page, key))
-
-def parse_rgb_string(value):
-    """Parse RGB values from string like '(0, 85, 255)' or similar formats.
-    
-    Args:
-        value (str): String containing RGB values.
-        
-    Returns:
-        tuple: RGB values as (r,g,b) tuple, or None if parsing fails.
-    """
-    if isinstance(value, str):
-        # Remove parentheses, brackets, etc.
-        clean_str = value.strip('()[]{}')
-        try:
-            # Split on commas and convert to integers
-            parts = [int(x.strip()) for x in clean_str.split(',')]
-            if len(parts) == 3:
-                return tuple(parts)
-        except:
-            pass
-    return None
